# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gathering Of The Fallen Telegram Mini App API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface Track {
  id: string;
  title: string;
  moods: string[];
  youtubeUrl: string;
  youtubeMusicUrl: string;
  spotifyUrl: string;
  appleMusicUrl: string;
}

export interface Release {
  id: string;
  title: string;
  description: string;
  date: string;
  url: string;
  accent: string;
}

export interface Video {
  id: string;
  title: string;
  duration: string;
  youtubeUrl: string;
  youtubeMusicUrl: string;
  thumbnailUrl: string;
}

export interface Experience {
  featured: Track;
  recentReleases: Release[];
  moods: string[];
  communityCount: number;
  trackCount: number;
}

export interface FanPost {
  id: number;
  author: string;
  message: string;
  createdAt: string;
}

export interface FanPostInput {
  /**
     * @minLength 2
     * @maxLength 700
     */
  message: string;
}

export interface FanPostModerationResult {
  id: number;
  hidden: boolean;
}

export interface AssistantInput {
  /**
     * @minLength 1
     * @maxLength 1800
     */
  message: string;
}

export interface AssistantReply {
  reply: string;
  suggestions: string[];
}

export interface SubscriptionInput {
  subscribed: boolean;
}

export interface SubscriptionState {
  subscribed: boolean;
}

export interface TelegramAuthInput {
  /**
     * @minLength 1
     * @maxLength 8192
     */
  initData: string;
}

export interface FanIdentity {
  id: string;
  displayName: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  isAdmin: boolean;
  subscribed: boolean;
}

export interface UsageSummary {
  todayRequests: number;
  monthRequests: number;
  monthEstimatedUsd: number;
  voiceSeconds: number;
  errorRate: number;
}

export type ListTracksParams = {
query?: string;
mood?: string;
};