# lib/api-zod/src/generated/types/fanPostModerationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gathering Of The Fallen Telegram Mini App API
 * OpenAPI spec version: 1.0.0
 */

export interface FanPostModerationResult {
  id: number;
  hidden: boolean;
}